backend/requirements.txt
flask 
flask-cors 
flask-sqlalchemy 
psycopg2-binary 
firebase-admin
python-dotenv
psycopg2-binary
requests
uagents>=0.22.0